Change coefficient type to integer in the fateman1 unpacked test.

+++ tests/fateman1_unpacked_perf.cpp
@@ -27,6 +27,7 @@
 
 #include "../src/environment.hpp"
 #include "../src/monomial.hpp"
+#include "../src/mp_integer.hpp"
 #include "../src/settings.hpp"
 
 using namespace piranha;
@@ -41,5 +42,5 @@ BOOST_AUTO_TEST_CASE(fateman1_unpacked_test)
 	if (boost::unit_test::framework::master_test_suite().argc > 1) {
 		settings::set_n_threads(boost::lexical_cast<unsigned>(boost::unit_test::framework::master_test_suite().argv[1u]));
 	}
-	BOOST_CHECK_EQUAL((fateman1<double,monomial<signed char>>().size()),135751u);
+	BOOST_CHECK_EQUAL((fateman1<integer,monomial<signed char>>().size()),135751u);
 }
